Fixed crash that happened with WW2GI. Thanks to supergoofy for helping me to find the problem.
The problem was there(GAME.CON)
// loading screen
onevent EVENT_GETLOADTILE
{
sound AAVE // uses spritesound and crashes
}
endevent

Note: the sound was supposed to be played once but with my approach it plays several sounds simultaneously and that's bad. I have chosen not to limit the sound command usage for WW2GI compatibility. I doubt whether I should commit this and decided that it's better to do after all.
If you have a better approach, you're welcome.