Re-add GAMETYPE_FRAGBAR to the default cases for deathmatch modes.
While the fragbar is indeed too cumbersome for use with more than 8 players, and a "show current score overlay" button is a better and more modern solution, this flag should connote whether a mode is *eligible* for it, not whether it is used at all.

TODO: Add a cvar for toggling the fragbar and turn it off by default.